+++ R/rubias.R
@@ -40,6 +40,7 @@
 #' @importFrom stats rbeta rmultinom var
 #' @importFrom utils write.table
 #' @importFrom Rcpp evalCpp
+#' @importFrom RcppParallel RcppParallelLibs
 #' @useDynLib rubias
 NULL